requirements.txt
psaw
youtube_dl
imgurpython==1.1.7
requests>=2.22.0
sqlalchemy>=1.3.8
praw>=6.0.0
lxml>=3.6.0
newspaper3k==0.2.8
beautifulsoup4==4.7.1
colorama==0.4.1
Pillow==6.2.0
pathvalidate==0.28.1
twitter-text-python==1.1.1
appdirs==1.4.3
alembic==1.3.1
flask==1.1.1
flask-socketio==4.2.1
eventlet==0.25.1
